+++ js/version/version_config.js
@@ -1,4 +1,4 @@
-export const current = "0.2.1";
+export const current = "0.2.0";
 export const version_check_path = "https://raw.githubusercontent.com/guinetn/wed/main/js/version_config.js";
 export const app_url = "https://github.com/guinetn/wed";